pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "argus"
version = "0.1.0"
description = "AI safety & evaluation research daily digest agent"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "groq>=0.11.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12.0",
    "mistune>=3.0.0",
]

[project.optional-dependencies]
reddit = ["praw>=7.7.0"]
all = ["argus[reddit]"]
dev = [
    "argus[all]",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
argus = "argus.runner:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["argus"]
